Incorporated updates to main.py file based on feedback

+++ main.py
@@ -22,39 +22,38 @@ jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir),
 #Retrieved from grc.com
 secret = "wGuA4zMzr465NsBeDSfpF5u3CC1LsD8JYJiVrsCOpWookykZzMVtV5BHX1g9Ta6"
 
-######   Globab render_str function   #####
-
 def render_str(template, **params):
+    """ Global render_str function """
     t = jinja_env.get_template(template)
     return t.render(params)
 
-######   Create a secure value   #####
 def create_secure_val(val):
+    """ Create a secure value """
     return val + "|" + hmac.new(secret, val).hexdigest()
 
-# Check a secure value
 def check_secure_val(secure_val):
+    """ Check a secure value """
     val = secure_val.split('|')[0]
     if secure_val == create_secure_val(val):
         return val
 
 #####   Password Security Methods   #####
 
-# Create random salt
 def create_salt():
+    """ Create random salt """
     return ''.join(random.SystemRandom().choice(
         string.ascii_uppercase + string.digits) for _ in range(9))    
 
-# Create password hash
 def create_pw_hash(email, pw, salt = None):
+    """ Create password hash """
     if not salt:
         salt = create_salt()
     h = hashlib.sha256(''.join([email, pw, salt])).hexdigest()
     return salt + "|" + h
 
-# Verify password hash
 def validate_pw(email, pw, h):
-    salt = h.split('|')[0]   # Should this be 0?
+    """ Verify password hash """
+    salt = h.split('|')[0]
     return h == create_pw_hash(email, pw, salt)
 
 ########################################################
@@ -63,8 +62,8 @@ def validate_pw(email, pw, h):
 
 ########################################################
 
-#MasterHandler Class
 class MasterHandler(webapp2.RequestHandler):
+    """ MasterHandler Class """
 
 #Jinja Methods
     def write(self, *a, **kw):
@@ -79,28 +78,28 @@ class MasterHandler(webapp2.RequestHandler):
 
 #Cookie Security Methods
 
-    # Set a secure cookie
     def set_secure_cookie(self, name, val):
+        """ Set a secure cookie """
         cookie_val = create_secure_val(val)
         self.response.headers.add_header(
             'Set-Cookie',
             '%s=%s; Path=/' % (name, cookie_val))
 
-    # Check a secure cookie
     def check_secure_cookie(self, name):
+        """ Check a secure cookie """
         cookie_val = self.request.cookies.get(name)
         return cookie_val and check_secure_val(cookie_val)
 
-    # Set cookie upon user login
     def login(self, user):
+        """ Set cookie upon user login """
         self.set_secure_cookie('user_id', str(user.key().id()))
 
-    # Reset cookie upon user logout
     def logout(self):
+        """ Reset cookie upon user logout """
         self.response.headers.add_header('Set-Cookie', 'user_id=; Path=/')
 
-    # Obtain user from cookie when pages are initialized
     def initialize(self, *a, **kw):
+        """ Obtain user from cookie when pages are initialized """
         webapp2.RequestHandler.initialize(self, *a, **kw)
         uid = self.check_secure_cookie('user_id')
         self.user = uid and User.by_id(int(uid))
@@ -111,30 +110,26 @@ class MasterHandler(webapp2.RequestHandler):
 
 ########################################################
 
-#Database to store user information
 class User(db.Model):
+    """ Database to store user information """
 
     email = db.EmailProperty(required=True)
     name = db.StringProperty(required=True)
     pw_hash = db.StringProperty(required=True)
 
-    #Get user by e-mail
     @classmethod
     def by_email(cls, email):
         return User.all().filter("email =", email).get()
 
-    #Get user by id
     @classmethod
     def by_id(cls, uid):
         return User.get_by_id(uid)
 
-    #Get user by name
     @classmethod
     def by_name(cls, uid):
         u = User.all().filter('name =', name).get()
         return u
 
-    #Create a new user object
     @classmethod
     def register(cls, email, name, pw):
         pw_hash = create_pw_hash(email, pw)
@@ -142,15 +137,14 @@ class User(db.Model):
                     name = name,
                     pw_hash = pw_hash)
 
-    #Login user
     @classmethod
     def login(cls, email, pw):
         u = cls.by_email(email)
         if u and validate_pw(email, pw, u.pw_hash):
             return u
 
-#Database to store blog post information
 class Post(db.Model):
+    """ Database to store blog post information """
 
     title = db.StringProperty(required=True)
     content = db.TextProperty(required=True)
@@ -161,12 +155,12 @@ class Post(db.Model):
     likes = db.IntegerProperty(default = 0)
 
     #Put line breaks in post content
-    def render(self):
+    def render(self):        
         self._render_text = self.content.replace('\n', '<br>')
         return render_str("post.html", p=self)
 
-#Database to store post comments
 class Comment(db.Model):
+    """ Database to store post comments """
 
     comment = db.StringProperty(required=True)
     post = db.StringProperty(required=True)
@@ -175,20 +169,20 @@ class Comment(db.Model):
     created = db.DateTimeProperty(auto_now_add=True)
     last_modified = db.DateTimeProperty(auto_now = True)
     
-    # Number of comments for post
+    #Retrieve total number of comments for post
     @classmethod
     def count_by_post_id(cls, post_id):
         c = Comment.all().filter('post =', post_id)
         return c.count()
 
-    # All comments for post
+    #Retrieve al comment for a post
     @classmethod
     def all_by_post_id(cls, post_id):
         c = Comment.all().filter('post =', post_id).order('-created')
         return c
 
-#Database to store post likes
 class Like(db.Model):
+    """ Database to store post likes """
     post_id = db.IntegerProperty(required = True)
     user_id = db.IntegerProperty(required=True)
 
@@ -202,6 +196,7 @@ class Like(db.Model):
 ##############    Front Page    #############
 
 class FrontPage(MasterHandler):
+    """ Front Page Handler """
     def get(self):
         # Retrieve all blog posts
         posts = db.GqlQuery("SELECT * FROM Post ORDER BY created DESC limit 10")
@@ -215,6 +210,7 @@ class FrontPage(MasterHandler):
 ##############    SignUp Page    #############
 
 class SignUpPage(MasterHandler):
+    """ Signup Page Handler """
     def get(self):
         self.render("signup.html", error=None)
 
@@ -253,6 +249,7 @@ class SignUpPage(MasterHandler):
 ##############    Login Page    #############
 
 class LoginPage(MasterHandler):
+    """ Login Page handler """
     def get(self):
         self.render("login.html", error=None)
 
@@ -272,6 +269,7 @@ class LoginPage(MasterHandler):
 ##############    Dashboard Page    #############
 
 class DashboardPage(MasterHandler):
+    """ Dashboard page handler """
     def get(self):
         # Retrieve all blog posts
         posts = db.GqlQuery("SELECT * FROM Post ORDER BY created DESC limit 10")
@@ -285,6 +283,7 @@ class DashboardPage(MasterHandler):
 ##############    Logout Page    #############
 
 class LogOutPage(MasterHandler):
+    """ Logout page handler """
     def get(self):
         self.logout()
         self.redirect("/")
@@ -292,8 +291,12 @@ class LogOutPage(MasterHandler):
 ##############    New Post Page    #############
 
 class NewPostPage(MasterHandler):
+    """ New Post page handler """
     def get(self):
-        self.render('newpost.html')
+        if self.user:
+            self.render('newpost.html')
+        else:
+            self.redirect('/login')
 
     def post(self):
         title = self.request.get('title')
@@ -313,6 +316,7 @@ class NewPostPage(MasterHandler):
 ##############    View Post Page    #############
 
 class ViewPostPage(MasterHandler):
+    """ View Post page handler """
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id))
         post = db.get(key)
@@ -347,6 +351,7 @@ class ViewPostPage(MasterHandler):
 ##############    Edit Post Page    #############
 
 class EditPostPage(MasterHandler):
+    """ Edit Post page handler """
     def get(self, post_id):
 
         # Retrieve all blog posts
@@ -367,26 +372,38 @@ class EditPostPage(MasterHandler):
     def post(self, post_id):
 
         if not self.user:
-            self.redirect('/dashboard')
+            self.redirect('/login')
+            return
 
         title = self.request.get('title')
         content = self.request.get('content')
 
-        if title and content:
-            key = db.Key.from_path('Post', int(post_id))
-            post = db.get(key)
-            post.title = title
-            post.content = content
-            post.put()
-            self.redirect('/post/%s' % post_id)
+        if post.user_id == self.user.key().id():
+            if title and content:
+                key = db.Key.from_path('Post', int(post_id))
+                post = db.get(key)
+                if post:
+                    post.title = title
+                    post.content = content
+                    post.put()
+                    self.redirect('/post/%s' % post_id)
+                else:
+                    error = "This post does not exist."
+                    self.render("dashboard.html", user = self.user.name,
+                    posts=posts, error=error)
+            else:
+                error = "You need both a title and some content to update a post."
+                self.render("editpost.html", title=title,
+                            content=content, error=error)
         else:
-            error = "You need both a title and some content to update a post."
-            self.render("editpost.html", title=title,
-                        content=content, error=error)
+                error = "You do not have access to edit this post."
+                self.render("dashboard.html", user = self.user.name,
+                    posts=posts, error=error)
 
 ##############    Delete Post Page    #############
 
 class DeletePostPage(MasterHandler):
+    """ Delete Post page handler """
     def get(self, post_id):
 
         # Retrieve all blog posts
@@ -395,12 +412,17 @@ class DeletePostPage(MasterHandler):
         if self.user:
             key = db.Key.from_path('Post', int(post_id))
             post = db.get(key)
-            if post.user_id == self.user.key().id():
-                post.delete()
-                error = "Your post has been deleted."
-                self.render("deletepost.html", error=error)
+            if post is not None:
+                if post.user_id == self.user.key().id():
+                    post.delete()
+                    error = "Your post has been deleted."
+                    self.render("deletepost.html", error=error)
+                else:
+                    error = "You do not have access to delete this post."
+                    self.render("dashboard.html", user = self.user.name,
+                        posts=posts, error=error)
             else:
-                error = "You do not have access to delete this post."
+                error = "This post does not exist."
                 self.render("dashboard.html", user = self.user.name,
                     posts=posts, error=error)
         else:
@@ -410,28 +432,33 @@ class DeletePostPage(MasterHandler):
 ##############    New Comment Page    #############
 
 class NewComment(MasterHandler):
+    """ New Comment handler """
     def post (self, post_id):
-        key = db.Key.from_path('Post', int(post_id))
-        post = db.get(key)
-        if not post:
-            self.error(404)
-            return
-        
-        comment = self.request.get('comment')
+        if self.user:
+            key = db.Key.from_path('Post', int(post_id))
+            post = db.get(key)
+            if not post:
+                self.error(404)
+                return
+            
+            comment = self.request.get('comment')
 
-        if comment:
-            c = Comment(comment=comment, post=post_id,
-                        user_id = self.user.key().id(),
-                        commentor = self.user.name)
-            c.put()
+            if comment:
+                c = Comment(comment=comment, post=post_id,
+                            user_id = self.user.key().id(),
+                            commentor = self.user.name)
+                c.put()
 
-            #A fix for datastore's Eventual Consistency
-            time.sleep(0.1)
-            self.redirect('/post/%s' % post_id)
+                #A fix for datastore's Eventual Consistency
+                time.sleep(0.1)
+                self.redirect('/post/%s' % post_id)
+        else:
+            self.redirect('/login')
 
 ##############    Edit Comment Page    #############
 
 class EditComment(MasterHandler):
+    """ Edit Comment handler """
     def get(self, post_id, comment_id):
         if self.user:
             post = Post.get_by_id(int(post_id))
@@ -455,28 +482,34 @@ class EditComment(MasterHandler):
         if self.user:
             post = Post.get_by_id(int(post_id))
             comment = Comment.get_by_id(int(comment_id))
+            if comment:
+                #Retrieve comment information
+                comments = Comment.all_by_post_id(post_id)
+                comments_count = Comment.count_by_post_id(post_id)
 
-            #Retrieve comment information
-            comments = Comment.all_by_post_id(post_id)
-            comments_count = Comment.count_by_post_id(post_id)
-
-            if comment.user_id == self.user.key().id():
-                comment_content = self.request.get("comment")
-                if comment_content:
-                    comment.comment = comment_content
-                    comment.put()
-                    time.sleep(0.1)
-                    self.redirect('/post/%s' % post_id)
-                else:
-                    error = "Please enter a comment."
-                    self.render(
-                        "editcomment.html",
+                if comment.user_id == self.user.key().id():
+                    comment_content = self.request.get("comment")
+                    if comment_content:
+                        comment.comment = comment_content
+                        comment.put()
+                        time.sleep(0.1)
+                        self.redirect('/post/%s' % post_id)
+                    else:
+                        error = "Please enter a comment."
+                        self.render(
+                            "editcomment.html",
                         comment=comment.comment)
+            else:
+                error = "This comment does not exist."
+                self.render("permalink.html", post=post,
+                        comments_count=comments_count,
+                        comments=comments, error=error)
         else:
             self.redirect("/login")
 
 ##############    Delete Comment    #############
 class DeleteComment(MasterHandler):
+    """ Delete Comment handler """
     def get(self, post_id, comment_id):
         if self.user:
             post = Post.get_by_id(int(post_id))
@@ -499,7 +532,8 @@ class DeleteComment(MasterHandler):
             self.redirect("/login")
 
 ##############    Likes Handler    #############
-class LikeComment(MasterHandler):
+class LikePost(MasterHandler):
+    """ Like Comment handler """
     def get(self, post_id):
         post = Post.get_by_id(int(post_id))
         user = self.user.key().id()
@@ -533,7 +567,7 @@ app = webapp2.WSGIApplication([
     ("/post/([0-9]+)", ViewPostPage),   
     ("/post/edit/([0-9]+)", EditPostPage),
     ("/post/delete/([0-9]+)", DeletePostPage),
-    ("/post/([0-9]+)/like", LikeComment),
+    ("/post/([0-9]+)/like", LikePost),
     ("/post/([0-9]+)/newcomment", NewComment),
     ("/post/([0-9]+)/comment/([0-9]+)/edit", EditComment),
     ("/post/([0-9]+)/comment/([0-9]+)/delete", DeleteComment)
